+++ src/components/themed/SimpleTextInput.tsx
@@ -1,11 +1,6 @@
 import * as React from 'react'
 import { useMemo } from 'react'
-import {
-  TextInput,
-  TouchableOpacity,
-  TouchableWithoutFeedback,
-  View
-} from 'react-native'
+import { TextInput, View } from 'react-native'
 import Animated, {
   interpolate,
   interpolateColor,
@@ -20,6 +15,8 @@ import Animated, {
 
 import { useHandler } from '../../hooks/useHandler'
 import { SpaceProps, useSpaceStyle } from '../../hooks/useSpaceStyle'
+import { EdgeTouchableOpacity } from '../common/EdgeTouchableOpacity'
+import { EdgeTouchableWithoutFeedback } from '../common/EdgeTouchableWithoutFeedback'
 import { styled, styledWithRef } from '../hoc/styled'
 import { AnimatedIconComponent, CloseIconAnimated } from '../icons/ThemedIcons'
 import { useTheme } from '../services/ThemeContext'
@@ -223,7 +220,7 @@ export const SimpleTextInput = React.forwardRef<
   )
 
   return (
-    <TouchableWithoutFeedback
+    <EdgeTouchableWithoutFeedback
       accessible={false}
       testID={testID}
       onPress={() => focus()}
@@ -278,7 +275,7 @@ export const SimpleTextInput = React.forwardRef<
           />
         </InnerContainer>
 
-        <TouchableOpacity
+        <EdgeTouchableOpacity
           accessible
           onPress={handleClearPress}
           testID={`${testID}.clearIcon`}
@@ -286,9 +283,9 @@ export const SimpleTextInput = React.forwardRef<
           <SideContainer size={rightIconSize}>
             <CloseIconAnimated color={iconColor} size={rightIconSize} />
           </SideContainer>
-        </TouchableOpacity>
+        </EdgeTouchableOpacity>
       </Container>
-    </TouchableWithoutFeedback>
+    </EdgeTouchableWithoutFeedback>
   )
 })