+++ src/components/ui4/ModalUi4.tsx
@@ -3,13 +3,7 @@
  */
 
 import * as React from 'react'
-import {
-  BackHandler,
-  Dimensions,
-  TouchableOpacity,
-  TouchableWithoutFeedback,
-  View
-} from 'react-native'
+import { BackHandler, Dimensions, View } from 'react-native'
 import { AirshipBridge } from 'react-native-airship'
 import {
   Gesture,
@@ -26,6 +20,8 @@ import Animated, {
 import AntDesignIcon from 'react-native-vector-icons/AntDesign'
 
 import { useHandler } from '../../hooks/useHandler'
+import { EdgeTouchableOpacity } from '../common/EdgeTouchableOpacity'
+import { EdgeTouchableWithoutFeedback } from '../common/EdgeTouchableWithoutFeedback'
 import { Theme, useTheme } from '../services/ThemeContext'
 import { EdgeText } from '../themed/EdgeText'
 import { BlurBackground } from './BlurBackground'
@@ -164,9 +160,9 @@ export function ModalUi4<T>(props: ModalPropsUi4<T>): JSX.Element {
 
   return (
     <>
-      <TouchableWithoutFeedback onPress={handleCancel}>
+      <EdgeTouchableWithoutFeedback onPress={handleCancel}>
         <Animated.View style={[styles.underlay, underlayStyle]} />
-      </TouchableWithoutFeedback>
+      </EdgeTouchableWithoutFeedback>
       <GestureDetector gesture={gesture}>
         <Animated.View style={[styles.modal, modalStyle, modalLayout]}>
           <BlurBackground />
@@ -185,7 +181,7 @@ export function ModalUi4<T>(props: ModalPropsUi4<T>): JSX.Element {
                 title ?? undefined
               )}
               {onCancel == null ? null : (
-                <TouchableOpacity
+                <EdgeTouchableOpacity
                   style={
                     isCustomTitle
                       ? styles.closeIconContainerAbsolute
@@ -198,7 +194,7 @@ export function ModalUi4<T>(props: ModalPropsUi4<T>): JSX.Element {
                     color={theme.deactivatedText}
                     size={theme.rem(1.25)}
                   />
-                </TouchableOpacity>
+                </EdgeTouchableOpacity>
               )}
             </View>
           )}